+++ actions.yaml
@@ -1,4 +1,4 @@
-# Copyright 2023 Canonical Ltd.
+# Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 get-certificate:
   description: Returns the TLS Certificate. Intended for testing and debugging purposes.
