+++ tests/integration/__init__.py
@@ -1,3 +1,3 @@
-# Copyright 2023 Canonical Ltd.
+# Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 """init file for the integration tests."""
