+++ tests/integration/any_charm_nginx_route.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Canonical Ltd.
+# Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
 # This Python script is designed to be loaded into any-charm. Some lint checks do not apply
