+++ tests/integration/test_cert_relation.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Canonical Ltd.
+# Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
 # pylint: disable=redefined-outer-name,unused-argument,duplicate-code
